added check for public/coverage already existing before putting in the ln

+++ lib/tasks/rspec.rake
@@ -38,7 +38,7 @@ namespace :spec do
     t.rcov_opts = lambda do
       IO.readlines("#{RAILS_ROOT}/spec/rcov.opts").map {|l| l.chomp.split " "}.flatten
     end
-    system "ln -s ../coverage public/coverage"
+    system "ln -s ../coverage public/coverage" unless File.exists?('public/coverage')
   end
   
   desc "Print Specdoc for all specs (excluding plugin specs)"
